File moved to Cryptography -> crc32 .h/cpp
