+++ src/shared/Cryptography/crc32.cpp
