+++ src/world/Server/Console/ConsoleCommands.cpp
@@ -9,7 +9,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Server/Console/BaseConsole.h"
 #include "Server/MainServerDefines.h"
 #include "Server/Master.h"
-#include "crc32.h"
+#include "Cryptography/crc32.h"
 #include "Server/World.h"
 #include "Management/ObjectMgr.h"
 #include "Server/Script/ScriptMgr.h"
